refactor: move style class to other package
