+++ inspector-runtime/src/main/java/com/uhufor/inspector/config/Config.kt
@@ -1,5 +1,6 @@
 package com.uhufor.inspector.config
 
+import com.uhufor.inspector.RelativeGuideStyle
 import com.uhufor.inspector.TraverseType
 import com.uhufor.inspector.UnitMode
 import kotlinx.coroutines.flow.MutableStateFlow
